refactor: Общий код вынесен в абстрактный класс
